+++ src/internal/connector/exchange/service_iterators.go
@@ -1,15 +1,15 @@
 package exchange
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
 	"fmt"
 	"strings"
 
 	multierror "github.com/hashicorp/go-multierror"
 	"github.com/microsoftgraph/msgraph-sdk-go/models"
-	msevents "github.com/microsoftgraph/msgraph-sdk-go/users/item/calendars/item/events"
-	cdelta "github.com/microsoftgraph/msgraph-sdk-go/users/item/contactfolders/item/contacts/delta"
-	mdelta "github.com/microsoftgraph/msgraph-sdk-go/users/item/mailfolders/item/messages/delta"
+	msuser "github.com/microsoftgraph/msgraph-sdk-go/users"
 	"github.com/pkg/errors"
 
 	"github.com/alcionai/corso/src/internal/connector/graph"
@@ -20,27 +20,50 @@ import (
 )
 
 const (
-	nextLinkKey  = "@odata.nextLink"
-	deltaLinkKey = "@odata.deltaLink"
+	metadataKey = "metadata"
 )
 
-// getAdditionalDataString gets a string value from the AdditionalData map. If
-// the value is not in the map returns an empty string.
-func getAdditionalDataString(
-	key string,
-	addtlData map[string]any,
-) string {
-	iface := addtlData[key]
-	if iface == nil {
-		return ""
+// makeMetadataCollection creates a metadata collection that has a file
+// containing all the delta tokens in tokens. Returns nil if the map does not
+// have any entries.
+//
+// TODO(ashmrtn): Expand this/break it out into multiple functions so that we
+// can also store map[container ID]->full container path in a file in the
+// metadata collection.
+func makeMetadataCollection(
+	tenant string,
+	user string,
+	cat path.CategoryType,
+	tokens map[string]string,
+	statusUpdater support.StatusUpdater,
+) (data.Collection, error) {
+	if len(tokens) == 0 {
+		return nil, nil
 	}
 
-	value, ok := iface.(*string)
-	if !ok {
-		return ""
+	buf := &bytes.Buffer{}
+	encoder := json.NewEncoder(buf)
+
+	if err := encoder.Encode(tokens); err != nil {
+		return nil, errors.Wrap(err, "serializing delta tokens")
 	}
 
-	return *value
+	p, err := path.Builder{}.ToServiceCategoryMetadataPath(
+		tenant,
+		user,
+		path.ExchangeService,
+		cat,
+		false,
+	)
+	if err != nil {
+		return nil, errors.Wrap(err, "making path")
+	}
+
+	return graph.NewMetadataCollection(
+		p,
+		[]graph.MetadataItem{graph.NewMetadataItem(graph.DeltaTokenFileName, buf.Bytes())},
+		statusUpdater,
+	), nil
 }
 
 // FilterContainersAndFillCollections is a utility function
@@ -59,6 +82,8 @@ func FilterContainersAndFillCollections(
 	var (
 		errs           error
 		collectionType = CategoryToOptionIdentifier(qp.Category)
+		// folder ID -> delta token for folder.
+		deltaTokens = map[string]string{}
 	)
 
 	for _, c := range resolver.Items() {
@@ -103,7 +128,7 @@ func FilterContainersAndFillCollections(
 			continue
 		}
 
-		jobs, _, err := fetchFunc(ctx, edc.service, qp.ResourceOwner, *c.GetId())
+		jobs, token, err := fetchFunc(ctx, edc.service, qp.ResourceOwner, *c.GetId())
 		if err != nil {
 			errs = support.WrapAndAppend(
 				qp.ResourceOwner,
@@ -113,6 +138,23 @@ func FilterContainersAndFillCollections(
 		}
 
 		edc.jobs = append(edc.jobs, jobs...)
+
+		if len(token) > 0 {
+			deltaTokens[*c.GetId()] = token
+		}
+	}
+
+	col, err := makeMetadataCollection(
+		qp.Credentials.AzureTenantID,
+		qp.ResourceOwner,
+		qp.Category,
+		deltaTokens,
+		statusUpdater,
+	)
+	if err != nil {
+		errs = support.WrapAndAppend("making metadata collection", err, errs)
+	} else if col != nil {
+		collections[metadataKey] = col
 	}
 
 	return errs
@@ -199,7 +241,7 @@ func FetchEventIDsFromCalendar(
 		ids  []string
 	)
 
-	options, err := optionsForCalendarEvents([]string{"id"})
+	options, err := optionsForEventsByCalendar([]string{"id"})
 	if err != nil {
 		return nil, "", err
 	}
@@ -234,7 +276,7 @@ func FetchEventIDsFromCalendar(
 			break
 		}
 
-		builder = msevents.NewEventsRequestBuilder(*nextLink, gs.Adapter())
+		builder = msuser.NewUsersItemCalendarsItemEventsRequestBuilder(*nextLink, gs.Adapter())
 	}
 
 	// Events don't have a delta endpoint so just return an empty string.
@@ -254,7 +296,7 @@ func FetchContactIDsFromDirectory(
 		deltaToken string
 	)
 
-	options, err := optionsForContactFoldersItem([]string{"parentFolderId"})
+	options, err := optionsForContactFoldersItemDelta([]string{"parentFolderId"})
 	if err != nil {
 		return nil, deltaToken, errors.Wrap(err, "getting query options")
 	}
@@ -266,8 +308,7 @@ func FetchContactIDsFromDirectory(
 		Delta()
 
 	for {
-		// TODO(ashmrtn): Update to pass options once graph SDK dependency is updated.
-		resp, err := sendContactsDeltaGet(ctx, builder, options, gs.Adapter())
+		resp, err := builder.Get(ctx, options)
 		if err != nil {
 			return nil, deltaToken, errors.Wrap(err, support.ConnectorStackErrorTrace(err))
 		}
@@ -286,19 +327,17 @@ func FetchContactIDsFromDirectory(
 			ids = append(ids, *item.GetId())
 		}
 
-		addtlData := resp.GetAdditionalData()
-
-		delta := getAdditionalDataString(deltaLinkKey, addtlData)
-		if len(delta) > 0 {
-			deltaToken = delta
+		delta := resp.GetOdataDeltaLink()
+		if delta != nil && len(*delta) > 0 {
+			deltaToken = *delta
 		}
 
-		nextLink := getAdditionalDataString(nextLinkKey, addtlData)
-		if len(nextLink) == 0 {
+		nextLink := resp.GetOdataNextLink()
+		if nextLink == nil || len(*nextLink) == 0 {
 			break
 		}
 
-		builder = cdelta.NewDeltaRequestBuilder(nextLink, gs.Adapter())
+		builder = msuser.NewUsersItemContactFoldersItemContactsDeltaRequestBuilder(*nextLink, gs.Adapter())
 	}
 
 	return ids, deltaToken, errs.ErrorOrNil()
@@ -317,7 +356,7 @@ func FetchMessageIDsFromDirectory(
 		deltaToken string
 	)
 
-	options, err := optionsForFolderMessages([]string{"id"})
+	options, err := optionsForFolderMessagesDelta([]string{"id"})
 	if err != nil {
 		return nil, deltaToken, errors.Wrap(err, "getting query options")
 	}
@@ -329,8 +368,7 @@ func FetchMessageIDsFromDirectory(
 		Delta()
 
 	for {
-		// TODO(ashmrtn): Update to pass options once graph SDK dependency is updated.
-		resp, err := sendMessagesDeltaGet(ctx, builder, options, gs.Adapter())
+		resp, err := builder.Get(ctx, options)
 		if err != nil {
 			return nil, deltaToken, errors.Wrap(err, support.ConnectorStackErrorTrace(err))
 		}
@@ -349,19 +387,17 @@ func FetchMessageIDsFromDirectory(
 			ids = append(ids, *item.GetId())
 		}
 
-		addtlData := resp.GetAdditionalData()
-
-		delta := getAdditionalDataString(deltaLinkKey, addtlData)
-		if len(delta) > 0 {
-			deltaToken = delta
+		delta := resp.GetOdataDeltaLink()
+		if delta != nil && len(*delta) > 0 {
+			deltaToken = *delta
 		}
 
-		nextLink := getAdditionalDataString(nextLinkKey, addtlData)
-		if len(nextLink) == 0 {
+		nextLink := resp.GetOdataNextLink()
+		if nextLink == nil || len(*nextLink) == 0 {
 			break
 		}
 
-		builder = mdelta.NewDeltaRequestBuilder(nextLink, gs.Adapter())
+		builder = msuser.NewUsersItemMailFoldersItemMessagesDeltaRequestBuilder(*nextLink, gs.Adapter())
 	}
 
 	return ids, deltaToken, errs.ErrorOrNil()