+++ src/internal/model/model.go
@@ -11,13 +11,17 @@ type (
 	Schema   int
 )
 
+// Schema constants denote the type of model stored. The integer values of the
+// constants can be changed without issue, but the string values should remain
+// the same. If the string values are changed, additional code will be needed to
+// transform from the old value to the new value.
+//
 //go:generate go run golang.org/x/tools/cmd/stringer -type=Schema
 const (
 	UnknownSchema = Schema(iota)
 	BackupOpSchema
 	RestoreOpSchema
 	BackupSchema
-	BackupDetailsSchema
 )
 
 // common tags for filtering
@@ -27,7 +31,7 @@ const (
 
 // Valid returns true if the ModelType value fits within the iota range.
 func (mt Schema) Valid() bool {
-	return mt > 0 && mt < BackupDetailsSchema+1
+	return mt > 0 && mt < BackupSchema+1
 }
 
 type Model interface {